+++ internal/repository/repository.go
@@ -7,8 +7,8 @@ import (
 )
 
 type SnippetRepository interface {
-	Insert(title, content string, expires int) (snippetID int64, err error)
-	Get(snippetID int64) (entities.Snippet, error)
+	Insert(title, content string, expires int) (snippetID int, err error)
+	Get(snippetID int) (entities.Snippet, error)
 	Latest(int) ([]entities.Snippet, error)
 }
